pyproject.toml
[project]
name = "days-since-last"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
]
